// src/api/api.spec.ts

import { AxiosResponse } from 'axios';
import mockAxios from '../__mocks__/axios';
import { fetchUsers, fetchRepos } from './api';
import {ApiResponse} from "../interfaces/ApiResponse.interface";
import GithubUser from "../models/github-user.model";
import GithubRepo from "../models/github-repo.model";

it('fetch user data from github search', async () => {
  mockAxios.get.mockImplementationOnce(() =>
    Promise.resolve<AxiosResponse<ApiResponse<GithubUser[]>>>({
      config: {},
      headers: {
        Accept: 'application/vnd.github.v3+json',
      },
      status: 200,
      statusText: 'cool',
      request: '',
      data: {
        incomplete_results: false,
        total_count: 1,
        items: [
          {
            login: 'anis',
            id: 1801091,
            node_id: 'MDQ6VXNlcjE4MDEwOTE=',
            avatar_url: 'https://avatars.githubusercontent.com/u/1801091?v=4',
            gravatar_id: '',
            url: 'https://api.github.com/users/anis',
            html_url: 'https://github.com/anis',
            followers_url: 'https://api.github.com/users/anis/followers',
            following_url:
              'https://api.github.com/users/anis/following{/other_user}',
            gists_url: 'https://api.github.com/users/anis/gists{/gist_id}',
            starred_url:
              'https://api.github.com/users/anis/starred{/owner}{/repo}',
            subscriptions_url:
              'https://api.github.com/users/anis/subscriptions',
            organizations_url: 'https://api.github.com/users/anis/orgs',
            repos_url: 'https://api.github.com/users/anis/repos',
            events_url: 'https://api.github.com/users/anis/events{/privacy}',
            received_events_url:
              'https://api.github.com/users/anis/received_events',
            type: 'User',
            site_admin: false,
            score: 1,
          },
        ],
      },
    })
  );

  const response = await fetchUsers('anis');
  const users = (await response.data.items) as GithubUser[];
  expect(
    users?.filter((e) => e.login.indexOf('anis') >= 0)?.length
  ).toBeGreaterThan(0);
  expect(mockAxios.get).toHaveBeenCalledTimes(1);
});

it('fetch repo data from github search', async () => {
  mockAxios.get.mockImplementationOnce(() =>
    Promise.resolve<AxiosResponse<ApiResponse<GithubRepo[]>>>({
      config: {},
      headers: {
        Accept: 'application/vnd.github.v3+json',
      },
      status: 200,
      statusText: 'cool',
      request: '',
      data: {
        incomplete_results: false,
        total_count: 1,
        items: [
          {
            id: 59824665,
            node_id: 'MDEwOlJlcG9zaXRvcnk1OTgyNDY2NQ==',
            name: 'hexo-theme-Anisina',
            full_name: 'Haojen/hexo-theme-Anisina',
            private: false,
            owner: {
              login: 'Haojen',
              id: 6928850,
              node_id: 'MDQ6VXNlcjY5Mjg4NTA=',
              avatar_url: 'https://avatars.githubusercontent.com/u/6928850?v=4',
              gravatar_id: '',
              url: 'https://api.github.com/users/Haojen',
              html_url: 'https://github.com/Haojen',
              followers_url: 'https://api.github.com/users/Haojen/followers',
              following_url:
                'https://api.github.com/users/Haojen/following{/other_user}',
              gists_url: 'https://api.github.com/users/Haojen/gists{/gist_id}',
              starred_url:
                'https://api.github.com/users/Haojen/starred{/owner}{/repo}',
              subscriptions_url:
                'https://api.github.com/users/Haojen/subscriptions',
              organizations_url: 'https://api.github.com/users/Haojen/orgs',
              repos_url: 'https://api.github.com/users/Haojen/repos',
              events_url:
                'https://api.github.com/users/Haojen/events{/privacy}',
              received_events_url:
                'https://api.github.com/users/Haojen/received_events',
              type: 'User',
              site_admin: false,
            },
            html_url: 'https://github.com/Haojen/hexo-theme-Anisina',
            description:
              ':art: A simple responsive , support qiniu image cdn theme for hexo https://haojen.github.io/ ',
            fork: false,
            url: 'https://api.github.com/repos/Haojen/hexo-theme-Anisina',
            forks_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/forks',
            keys_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/keys{/key_id}',
            collaborators_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/collaborators{/collaborator}',
            teams_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/teams',
            hooks_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/hooks',
            issue_events_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/issues/events{/number}',
            events_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/events',
            assignees_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/assignees{/user}',
            branches_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/branches{/branch}',
            tags_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/tags',
            blobs_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/git/blobs{/sha}',
            git_tags_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/git/tags{/sha}',
            git_refs_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/git/refs{/sha}',
            trees_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/git/trees{/sha}',
            statuses_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/statuses/{sha}',
            languages_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/languages',
            stargazers_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/stargazers',
            contributors_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/contributors',
            subscribers_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/subscribers',
            subscription_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/subscription',
            commits_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/commits{/sha}',
            git_commits_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/git/commits{/sha}',
            comments_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/comments{/number}',
            issue_comment_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/issues/comments{/number}',
            contents_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/contents/{+path}',
            compare_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/compare/{base}...{head}',
            merges_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/merges',
            archive_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/{archive_format}{/ref}',
            downloads_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/downloads',
            issues_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/issues{/number}',
            pulls_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/pulls{/number}',
            milestones_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/milestones{/number}',
            notifications_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/notifications{?since,all,participating}',
            labels_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/labels{/name}',
            releases_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/releases{/id}',
            deployments_url:
              'https://api.github.com/repos/Haojen/hexo-theme-Anisina/deployments',
            created_at: '2016-05-27T09:57:08Z',
            updated_at: '2021-01-27T13:02:33Z',
            pushed_at: '2020-12-29T04:25:45Z',
            git_url: 'git://github.com/Haojen/hexo-theme-Anisina.git',
            ssh_url: 'git@github.com:Haojen/hexo-theme-Anisina.git',
            clone_url: 'https://github.com/Haojen/hexo-theme-Anisina.git',
            svn_url: 'https://github.com/Haojen/hexo-theme-Anisina',
            homepage: '',
            size: 2296,
            stargazers_count: 746,
            watchers_count: 746,
            language: 'HTML',
            has_issues: true,
            has_projects: true,
            has_downloads: true,
            has_wiki: true,
            has_pages: false,
            forks_count: 159,
            mirror_url: null,
            archived: false,
            disabled: false,
            open_issues_count: 12,
            license: {
              key: 'apache-2.0',
              name: 'Apache License 2.0',
              spdx_id: 'Apache-2.0',
              url: 'https://api.github.com/licenses/apache-2.0',
              node_id: 'MDc6TGljZW5zZTI=',
            },
            forks: 159,
            open_issues: 12,
            watchers: 746,
            default_branch: 'master',
            score: 1,
          },
        ],
      },
    })
  );

  const response = await fetchRepos('theme');
  const users = (await response.data.items) as GithubRepo[];
  expect(
    users?.filter((e) => e.name.indexOf('theme') >= 0)?.length
  ).toBeGreaterThan(0);
  expect(mockAxios.get).toHaveBeenCalledTimes(1);
});
